==> 1. imdb_episode_scraper.ipynb <==
The notebook loops through each season of the Simpsons on IMDb and pulls the basic information provided.

As I want to keep the number of hits on IMDb to a reasonable amount, the loop only pulls data from the season tables, and does not drill down into each episode (despite that offering more data).

# Simpsons IMDb link
# https://www.imdb.com/title/tt0096697/episodes?season=1

from requests import get
from bs4 import BeautifulSoup
import pandas as pd

The fields found on the webpage correspond to the descriptions given below

# Below are outdated

#episode_containers[0].a['title']) # Title
#episode_containers[0].meta['content']) # Episode Number
#episode_containers[0].find('div', class_='airdate').text.strip() # Air Date
#episode_containers[0].find('span', class_='ipl-rating-star__rating').text # Episode Rating
#episode_containers[0].find('span', class_='ipl-rating-star__total-votes').text # Total No. Of Votes
#episode_containers[0].find('div', class_='item_description').text.strip() # Episode Description

# Defining Functions

def get_season_url(season):
    """Return the URL for a chosen season"""
    simpsons_url = 'https://www.imdb.com/title/tt0096697/episodes?season='
    season_url = simpsons_url + str(season)
    return season_url

def extract_season_data(season_url):
    """Extract the webpage data and pull out each individual episode into an 'episode container' which contains its
    row in the table."""
        
    # Request from the server the content of the web page by using get()
    response = get(season_url, headers={'User-Agent':'Mozilla/5.0'})

    # Parse the content of the request with BeautifulSoup
    page_html = BeautifulSoup(response.text, 'html.parser')

    # Select all the episode containers from the season's page
    episode_containers = page_html.find_all('article', class_='episode-item-wrapper')
    
    return episode_containers

def extract_ep_data_from_container(episode_container, season):
    """Extract the individual field data from the 'episode container' of a single episode."""
    
    try:
        name = episode_container.find('div', class_='ipc-title__text').text.strip()
        title_divider = name.index('∙')
        episode_number = name[:title_divider]
        ep_index = episode_number.index('.')
        episode_number = episode_number[ep_index+2:]
        title = name[title_divider+1:]
    except:
        title = ''
        episode_number = ''

    try:
        airdate = episode_container.find('span', class_='jAfkDE').text.strip()
    except:
        airdate = ''

    try:
        rating = episode_container.find('span', class_='ipc-rating-star').text.strip()
        rating_index = rating.index('/')
        rating = rating[:rating_index]
    except:
        rating = ''

    try:
        total_votes = episode_container.find('span', class_='ipc-rating-star--voteCount').text
        total_votes = total_votes.replace('(','')
        total_votes = total_votes.replace(')','')
    except:
        total_votes = ''

    try:
        desc = episode_container.find('div', class_='ipc-html-content-inner-div').text.strip()
    except:
        desc = ''
        
    # Compiling the episode info
    episode_data = [season, episode_number, title, airdate, rating, total_votes, desc]
    
    return episode_data

def convert_ep_array_to_dataframe(episodes):
    """Convert the array of episode data into a pandas dataframe"""
    episodes = pd.DataFrame(episodes, columns=['season', 'episode_no', 'title', 
                                               'airdate', 'rating', 'total_votes', 'desc'])
    return episodes  

def multiply_votes(vote_str):
    """Convert the total_votes extracted from IMDB into full numbers (remove the use of K 
    to indicate thousands)"""
    try:
        index = vote_str.index('K')
    except:
        index = None
        
    if index:
        vote_str = vote_str[:index]
        vote = float(vote_str)*1000
    else:
        vote = float(vote_str)
        
    return int(vote)

def gather_imdb_data(max_season = 33):
    """Gather data from IMDb for episodes of the simpsons"""
    # Initializing the series that the loop will populate
    ep_array = []
    
    # For every season in the series-- range depends on the show
    for sn in range(1, max_season+1):
        season_url = get_season_url(sn)
        episode_containers = extract_season_data(season_url)

        # For each episode in each season
        for episode in episode_containers:
            episode_data = extract_ep_data_from_container(episode, sn)

            # Append the episode info to the complete dataset
            ep_array.append(episode_data)

    ep_dataframe = convert_ep_array_to_dataframe(ep_array)
    
    ep_dataframe['total_votes'] = ep_dataframe['total_votes'].apply(multiply_votes)
    ep_dataframe['airdate'] = ep_dataframe['airdate'].apply(lambda x: datetime.strptime(x, "%a, %b %d, %Y"))
    
    return ep_dataframe

# Gather Data

simpsons_episodes = gather_imdb_data()

simpsons_episodes.head()

simpsons_episodes.to_pickle("./data/simpsons_episodes.pkl")

print('Data was extracted for {} episodes.'.format(simpsons_episodes.shape[0]))

simpsons_episodes
