==> 2. wiki_viewer_scraper.ipynb <==
This notebook pulls data from wikipedia, which includes the number of viewers each episode had on its original airing (within the US).

# First 20 seasons
# https://en.wikipedia.org/wiki/List_of_The_Simpsons_episodes_(seasons_1%E2%80%9320)

# Seasons 21 - current
# https://en.wikipedia.org/wiki/List_of_The_Simpsons_episodes_(season_21%E2%80%93present)

from requests import get
from bs4 import BeautifulSoup
import pandas as pd
import re

# Define Functions

def scrape_ep_tables(ep_url):
    """Get episode tables (for each season) from wikipedia"""
    
    # Request from the server the content of the web page by using get()
    response = get(ep_url)

    # Parse the content of the request with BeautifulSoup
    page_html = BeautifulSoup(response.text, 'html.parser')

    # Extract each season table into ep_tables. Each index is a seperate season
    ep_tables = page_html.find_all('table', class_='wikiepisodetable')
    
    return ep_tables

def extract_row_cell(cells, row_span, row_cont):
    """Extract values from a row of wikitable (accounting for rowspan)"""
    row = []
    column_nm = ['ep_no', 'title', 'directed', 'written', 'airdate', 'prod_code', 'viewers']
    
    i = 0
    for key in column_nm:
        rs = row_span[key]
        rc = row_cont[key]
        
        if rs > 0:
            value = '' #rc
            rs += -1
        else:
            value = cells[i].text.strip()
            if cells[i].has_attr("rowspan"):
                rs = int(cells[i]["rowspan"]) - 1
                rc = value
                
            i += 1

        if rs == 0:
                rc = ''
                
        row.append(value)
        row_span[key] = rs
        row_cont[key] = rc
        
    return row, row_span, row_cont

def extract_ep_details(rows, season):
    """Extract Details of Episodes"""
    
    season_list = []
    rowspan = {'ep_no': 0, 'title': 0, 'directed': 0, 'written': 0, 'airdate': 0, 'prod_code': 0, 'viewers': 0, 'overall_ep': 0}
    row_content = {'ep_no': '', 'title': '', 'directed': '', 'written': '', 'airdate': '', 'prod_code': '', 'viewers': '', 'overall_ep': ''}
    column_nm = ['ep_no', 'title', 'directed', 'written', 'airdate', 'prod_code', 'viewers']
    
    for row in rows:
        ep_header = row.find('th')
        if (ep_header):
            # Update overall ep no if th present
            overall_ep = ep_header.text.strip()

        cols=row.find_all('td')
        if len(cols) > 0: # Account for seemingly empty rows
            line, rowspan, row_content = extract_row_cell(cols, rowspan, row_content)
            line.append(overall_ep) # Append the overall episode number
            line.append(season) # Append Season No
            season_list.append(line)
        
    simpsons_episodes = pd.DataFrame(season_list, columns=['episode', 'ep_title', 'director', 'writer', 
                                                   'air_date', 'prod_code', 'viewers', 'ep_overall', 'season'])  
    return simpsons_episodes

def identify_exceptions(rows):
    """Identify tables which have been found but do not represent a season of episodes (e.g. movie)"""
    skip_flag = False
    
    headers = rows[0].find_all('th')
    for header in headers:
        header = header.text.strip()
        if 'Release date' in header:
            skip_flag = True
            
    return skip_flag

def trim_viewers(row):
    index = row.find('[')
    viewers = row[:index]
    return viewers

def remove_reference_tags(ep_list):
    """Use regular expressions to identify reference tags and remove them"""
    pattern = r'\[[^\]]*\]'
    index = ep_list['viewers'].find('[')
    ep_list['viewers']
    ep_list['viewers'] = ep_list['viewers'].apply(lambda x: re.sub(pattern, '', x))
    return ep_list

def remove_quotes(ep_list):
    """Remove quotations from the episode titles"""

    ep_list['ep_title'] = ep_list['ep_title'].apply(lambda x: re.sub('"', '', x))
    return ep_list

def generate_episode_dataframe(ep_tables, max_season = 33):

    ep_list = pd.DataFrame(columns=['episode', 'ep_title', 'director', 'writer', 
                                    'air_date', 'prod_code', 'viewers', 'ep_overall', 'season'])
    season = 1

    for i in range(len(ep_tables)):
        if season >= max_season+1:
            continue

        season_table = ep_tables[i]
        season_rows = season_table.find_all('tr')

        if identify_exceptions(season_rows): # Skip table if not a season of episodes
            continue

        simpsons_episodes = extract_ep_details(season_rows, season)
        ep_list = pd.concat([ep_list, simpsons_episodes])
        season += 1 # Increment season if table not skipped
        
    #ep_list = remove_reference_tags(ep_list)
    ep_list['viewers'] = ep_list['viewers'].apply(lambda x: trim_viewers(x))
    ep_list = remove_quotes(ep_list)
    
    ep_list.reset_index(drop=True, inplace=True)
        
    return ep_list

# Build Dataset

url1 = 'https://en.wikipedia.org/wiki/List_of_The_Simpsons_episodes_(seasons_1%E2%80%9320)'
url2 = 'https://en.wikipedia.org/wiki/List_of_The_Simpsons_episodes_(season_21%E2%80%93present)'

ep_tables1 = scrape_ep_tables(url1)
ep_tables2 = scrape_ep_tables(url2)

ep_tables = ep_tables1 + ep_tables2

ep_list = generate_episode_dataframe(ep_tables)

ep_list.head(30)

ep_list.shape

# Save Data

ep_list.to_pickle("./data/simpsons_wiki_episodes.pkl")
