==> 3. guest_star_scraper.ipynb <==
# Celebrity Guest Stars (from Wikipedia)

Identify which episodes had Special Guest Stars. 

Unfortunately, there does not appear to be a simple list for this, so we have to use the full list of guest stars and reduce it.

# First 20 seasons
# https://en.wikipedia.org/wiki/List_of_The_Simpsons_guest_stars_(seasons_1%E2%80%9320)

# Seasons 21 - current
# https://en.wikipedia.org/wiki/List_of_The_Simpsons_guest_stars_(seasons_21%E2%80%93present)

from requests import get
from bs4 import BeautifulSoup
import pandas as pd

# Functions

def scrape_tables(wiki_url):
    """Get episode tables (for each season) from wikipedia"""
    
    # Request from the server the content of the web page by using get()
    response = get(wiki_url)

    # Parse the content of the request with BeautifulSoup
    page_html = BeautifulSoup(response.text, 'html.parser')
    guest_table = page_html.find('table', class_='sortable')
    
    return guest_table

def extract_row_details(rows):
    guest_list = []
    for row in rows:
        name = row.find('th')
        name = name.text.strip()
        cols=row.find_all('td')
        cols=[x.text.strip() for x in cols]
        cols.append(name)
        guest_list.append(cols)

    del guest_list[0] # Delete header 
    return guest_list

def convert_row_array(guest_list):
    guests = pd.DataFrame(guest_list, columns=['season', 'role', 'no', 
                                           'prodCode', 'epTitle', 'GuestStar'])
    return guests

# Scrape wikipedia

url1 = 'https://en.wikipedia.org/wiki/List_of_The_Simpsons_guest_stars_(seasons_1%E2%80%9320)'
url2 = 'https://en.wikipedia.org/wiki/List_of_The_Simpsons_guest_stars_(seasons_21%E2%80%93present)'

guest_tables1 = scrape_tables(url1)
guest_tables2 = scrape_tables(url2)

rows1 = guest_tables1.find_all('tr')
rows2 = guest_tables2.find_all('tr')
rows = rows1 + rows2

guest_list = extract_row_details(rows)

guests = convert_row_array(guest_list)
guests = guests.dropna()

guests.head()

guests.shape

guests['GuestStar'].value_counts()

#guests.to_excel('data/guest_stars.xlsx')

# Filtering out the 'Celebrity Guests'

### Limited Appearances

Guests that only appeared up to 5 times (some actors, such as JK Simmons have appeared 3 times).

ep_count = guests['GuestStar'].value_counts()
keep_names1 = ep_count[ep_count == 1].index # Idenfity people that appeared once
keep_names2 = ep_count[ep_count == 2].index # Idenfity people that appeared once
keep_names3 = ep_count[ep_count == 3].index # Idenfity people that appeared once
keep_names4 = ep_count[ep_count == 4].index # Idenfity people that appeared once
keep_names5 = ep_count[ep_count == 5].index # Idenfity people that appeared once

keep_names = keep_names1.append(keep_names2).append(keep_names3).append(keep_names4).append(keep_names5)

guests_single_appearance = guests[guests['GuestStar'].isin(keep_names)]

guests_single_appearance.head()

guests_single_appearance.shape

### Starring as themselves

Identify guests that are listed as 'Himself', 'Herself', 'Themselves', 'Narrator'

celeb_markers = ['Himself', 'Herself', 'Themselves', 'Narrator'] # People credited as themselves are most likely celebrity guests

guests_themselves = pd.DataFrame(columns = guests.columns)
for role in celeb_markers:
    temp = guests[guests['role'].str.contains(role)]
    guests_themselves = pd.concat([guests_themselves, temp])

guests_themselves.head()

### Remove starring as themselves from limited appearances

We do not to manually go through any guest stars that are listed as 'themselves' etc. so we can remove these from the guests to be manually parsed.

a_index = guests_single_appearance.index
b_index = guests_themselves.index
mask = ~a_index.isin(b_index)
guests_single_appearance = guests_single_appearance.loc[mask]

guests_single_appearance.shape

guests_single_appearance.to_excel('data/guest_stars_limited.xlsx')

**At this point,the `guest_stars_limited` excel should be manually parsed to only include 'real' celebrity guests.**

- Kept no one that has 'archival' footage
- Has a wikipedia page
- Known primarily for non voice acting work
- Not a writer on the simpsons

# Combine into a single data-frame

guests_themselves['As_Themselves'] = 1

# Reload the hand filtered list
df_hand_filter = pd.read_excel('data/guest_stars_hand_filtered.xlsx', index_col=0)  
df_hand_filter['Character_Played'] = 1

df_hand_filter.head()

guests_list = pd.concat([guests_themselves, df_hand_filter])
guests_list = guests_list.reset_index(drop=True)
guests_list = guests_list.drop_duplicates()

guests_list.shape

guests_list.As_Themselves = guests_list.As_Themselves.fillna(0)
guests_list.Character_Played = guests_list.Character_Played.fillna(0)

guests_list.head()

# Save curated Guest List

#guests_list.to_excel('data/guest_stars.xlsx')

guests_list.to_pickle("./data/simpsons_guest_stars.pkl")
