==> 4. episode_preprocessing.ipynb <==
import pandas as pd

# Load all Data

def combine_rows(x):
    """Join together multiple rows worth of data for a single column."""
    x = x.to_list()
    output = ' & '.join(x)
    return output

## IMDB data

simpsons_episodes = pd.read_pickle("./data/simpsons_episodes.pkl")

simpsons_episodes[['season', 'episode_no', 'total_votes']] = simpsons_episodes[['season', 'episode_no', 'total_votes']].astype('int')

simpsons_episodes.head()

## Wikipedia data

simpsons_wiki = pd.read_pickle("./data/simpsons_wiki_episodes.pkl")

simpsons_wiki[['season', 'ep_overall']] = simpsons_wiki[['season', 'ep_overall']].astype('int')

simpsons_wiki.head()

agg_functions = {'episode': 'first', 'ep_title': 'first', 'director': combine_rows, 'writer': combine_rows, 'air_date': 'first', 
                 'prod_code': 'first', 'viewers': 'first', 'ep_overall': 'first', 'season': 'first'}
simpsons_wiki_group = simpsons_wiki.groupby(simpsons_wiki['ep_overall']).aggregate(agg_functions)

simpsons_wiki_group[['episode']] = simpsons_wiki_group[['episode']].astype('int')

simpsons_wiki_group.head()

## Guest Appearances

simpsons_guests = pd.read_pickle("./data/simpsons_guest_stars.pkl")

simpsons_guests = simpsons_guests[simpsons_guests.season != 'Movie']

simpsons_guests[['season', 'As_Themselves', 'Character_Played']] = simpsons_guests[['season', 'As_Themselves', 'Character_Played']].astype('int')

simpsons_guests.head()

agg_functions = {'season': 'first', 'prodCode': 'first', 'epTitle': 'first', 'As_Themselves': 'sum', 'Character_Played': 'sum'}
simpsons_guests_slim = simpsons_guests.groupby(simpsons_guests['prodCode']).aggregate(agg_functions)
simpsons_guests_slim.reset_index(drop = True, inplace=True)

simpsons_guests_slim.head()

simpsons_guests_slim.shape

# Combine data

simpsons_full = simpsons_episodes.merge(simpsons_wiki_group, left_on = ['season', 'episode_no'], right_on = ['season', 'episode'])

simpsons_full.drop(['ep_title', 'air_date', 'episode'], axis = 1, inplace = True)

simpsons_full = simpsons_full.merge(simpsons_guests_slim, left_on = 'prod_code', right_on = ['prodCode'], how = 'left')

simpsons_full.drop(['prodCode', 'epTitle', 'season_y'], axis = 1, inplace = True)

simpsons_full.rename(columns={'season_x': 'season', 'episode_no': 'episode'},
                    inplace = True)

simpsons_full['As_Themselves'] = simpsons_full['As_Themselves'].fillna(0)
simpsons_full['Character_Played'] = simpsons_full['Character_Played'].fillna(0)

simpsons_full[['As_Themselves', 'Character_Played']] = simpsons_full[['As_Themselves', 'Character_Played']].astype('int')

# Round total_votes to nearest hundred as this is what IMDb automatically does to values in the thousands
simpsons_full['total_votes'] = simpsons_full['total_votes'].apply(lambda x: round(x, -2))

# Feature Engineering

import re

def split_dir_writer(field):
    """Split the director and writer fields into a list instead of one string"""
    reg_pattern = r' and |, | & '
    output = re.split(reg_pattern, field)
    
    # Deduplicate in case a writer worked on multiple stories in one episode (Treehouse of Horror)
    output = list(set(output))
    return output

def is_main_character(row, character_str):
    """Check whether or not a character is one of the 'main' characters in an episode. This is decided by whether an episode title or description
    contains their name. """
    flag1 = character_str in row['title']
    flag2 = character_str in row['desc']
    
    if flag1 or flag2:
        return 1
    else:
        return 0

# Highlight episodes were a member of the family is a 'main' character (according to the title or episode description)
simpsons_full['Homer_main'] = simpsons_full.apply(lambda x: is_main_character(x, 'Homer'), axis = 1)
simpsons_full['Marge_main'] = simpsons_full.apply(lambda x: is_main_character(x, 'Marge'), axis = 1)
simpsons_full['Bart_main'] = simpsons_full.apply(lambda x: is_main_character(x, 'Bart'), axis = 1)
simpsons_full['Lisa_main'] = simpsons_full.apply(lambda x: is_main_character(x, 'Lisa'), axis = 1)
simpsons_full['Maggie_main'] = simpsons_full.apply(lambda x: is_main_character(x, 'Maggie'), axis = 1)

# Convert the director and writer fields into a list (seperate on commas, ampersands and 'and')
simpsons_full['director'] = simpsons_full.apply(lambda x: split_dir_writer(x['director']), axis = 1)
simpsons_full['writer'] = simpsons_full.apply(lambda x: split_dir_writer(x['writer']), axis = 1)

# Flag the Treehouse of Horror episodes (create a flag to identify them)
simpsons_full['Treehouse'] = simpsons_full['title'].apply(lambda x: 1 if 'Treehouse' in x else 0)

# Save Data

simpsons_full.head()

simpsons_full.shape

simpsons_full.to_pickle("./data/simpsons_processed.pkl")
